requirements.txt
PyYAML
playwright
openai
anthropic
google-generativeai
